Add iQueSignatureDecrypt, gets the expected hash from a signature if we have the public key
Unfortunately requires ManagedOpenSsl as there doesn't seem to be any public methods in .NET for decrypting signatures, kinda silly when they have to decrypt the signature to verify the hash in the first place...
(you'll have to copy the x86\libeay32.dll & x86\ssleay32.dll to an x86 folder next to iquetool.exe, otherwise signature decryption / expected hash will be disabled)

Structs containing signatures will also have the expected hash shown when printing info & will also print a big "contact me!" message if any hashes are found starting with 0, as that might let us try fakesigning again!
(last fakesign method replaced the signature with all 00, meaning it had bad PKCS1 padding - might be why it failed. If we can find a decrypted sig that starts with 00 it'll be completely valid & usable for us to fakesign with)
I've checked the tickets from ~10 or so NANDs and none seem to have hashes starting with 00.. but now more people are starting to dump their units we should hopefully find one eventually.